requirements.txt
channels==3.0.3
Django==3.1.5
django-admin-interface==0.19.1
django-model-utils==4.2.0
djangorestframework==3.13.1
django-rest-swagger==2.2.0
drf-yasg==1.21.3
PyYAML==6.0